staterecovery: carveout executeCommand fn to it's own file
