+++ coordinator/ethereum/test-utils/src/main/kotlin/net/consensys/zkevm/ethereum/MakefileContractDeploymentHelper.kt
@@ -1,78 +1,11 @@
 package net.consensys.zkevm.ethereum
 
 import build.linea.contract.l1.LineaContractVersion
-import net.consensys.linea.async.toSafeFuture
-import net.consensys.linea.testing.filesystem.getPathTo
-import org.apache.logging.log4j.LogManager
+import linea.testing.CommandResult
+import linea.testing.Runner
 import tech.pegasys.teku.infrastructure.async.SafeFuture
-import java.io.BufferedReader
-import java.io.File
-import java.io.InputStreamReader
 import java.util.regex.Matcher
 import java.util.regex.Pattern
-
-data class CommandResult(
-  val exitCode: Int,
-  val stdOut: List<String>,
-  val stdErr: List<String>
-)
-
-fun executeCommand(
-  command: String,
-  envVars: Map<String, String> = emptyMap(),
-  executionDir: File = getPathTo("Makefile").parent.toFile()
-): SafeFuture<CommandResult> {
-  val log = LogManager.getLogger("net.consensys.zkevm.ethereum.CommandExecutor")
-  val processBuilder = ProcessBuilder("/bin/sh", "-c", command)
-  processBuilder.directory(executionDir)
-
-  // Set environment variables
-  val env = processBuilder.environment()
-  for ((key, value) in envVars) {
-    env[key] = value
-  }
-
-  val process = processBuilder.start()
-  val stdOutReader = BufferedReader(InputStreamReader(process.inputStream))
-  val stdErrorReader = BufferedReader(InputStreamReader(process.errorStream))
-
-  // Read the standard output
-  log.debug(
-    "going to execute command: dir='{}', command='{}', envVars={} commandProcessId={} processInfo={}",
-    executionDir,
-    command,
-    envVars,
-    process.pid(),
-    process.info()
-  )
-  process.waitFor(60, java.util.concurrent.TimeUnit.SECONDS)
-  val futureResult = process
-    .onExit()
-    .thenApply { processResult ->
-      val stdOutLines = stdOutReader.lines().toList()
-      val stdErrLines = stdErrorReader.lines().toList()
-      log.debug(
-        "command finished: dir='{}', command='{}', exitCode={} envVars={} processId={} threadId={}",
-        executionDir,
-        command,
-        processResult.exitValue(),
-        envVars,
-        ProcessHandle.current().pid(),
-        Thread.currentThread().threadId()
-      )
-      log.debug(
-        "stdout: {}",
-        stdOutLines.joinToString("\n")
-      )
-      log.debug(
-        "stderr: {}",
-        stdErrLines.joinToString("\n")
-      )
-      CommandResult(processResult.exitValue(), stdOutLines, stdErrLines)
-    }
-
-  return futureResult.toSafeFuture()
-}
 
 internal val lineaRollupAddressPattern = Pattern.compile(
   "^contract=LineaRollup(?:.*)? deployed: address=(0x[0-9a-fA-F]{40}) blockNumber=(\\d+)"
@@ -112,7 +45,7 @@ private fun deployContract(
   env: Map<String, String> = emptyMap(),
   addressPattern: Pattern
 ): SafeFuture<DeployedContract> {
-  return executeCommand(
+  return Runner.executeCommand(
     command = command,
     envVars = env
   )
